Add random solution generator - with constraints

+++ src/models/entities/base.py
@@ -1,10 +1,10 @@
-from typing import NamedTuple
+from typing import NamedTuple, Optional
 
 
 class BaseEntity(NamedTuple):
-    index: int
-    max_capacity: int
-    startup_cost: int
+    index: int  # ! counted starting from from "1"
+    max_capacity: float
+    startup_cost: float
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__} ({self.index}): cap ({self.max_capacity}), startup ({self.startup_cost})"
